requirements.txt
flask==3.0.0
torch==2.11.0
torchvision==0.26.0
tensorflow==2.21.0
pillow==12.2.0
numpy==2.4.4
gunicorn==21.2.0
